First PoC to test the Architecture Overview
This commit presents a simple PoC to test the architecture that is
wanted to be used by the cyclone project.
Only basic functionalities are introduced and some hacks are required
to proof the architecture. However serves as good starting point.

One of the most interesting points is that, if correctly configured,
this PoC could actually deploy the VPNs, LBs and Firewalls required
by the project. They wouyld be deployed in a dirty and not-so-robust
way, but they would work