Present the first skeleton of the Client VPN App
A first attempt to build the skeleton of the client app is presented.
It is very similar to the server app.

The idea is to spawn a rest server that expects files uploaded from the
clients, and save those files into a working dir.
After that, the client is expected to launch a couple bash commands
that build or run the VPN Client Docker app, given the previous files
as input