Refactor Imports and Order some Methods
For somereason there are lots of imports that start by main.python
instead of src.main.python. All the import statements should be from
the same source in order to easy the way the python paths should be
configured.

On the other side, some methods have been moved to increase the
readability of the code and promoto a cleaner way to extend the
files or classes.