Add Test for the Configurator App
Since all the deployed services are distributed, there is the need to
check wether they are or will be stable after some remode modules use
them.

In the case of the configurator, there is the need to check that the
rest server responds as expected when is deployed and when the
ServiceMaker class will create a REST client on the Fly.

This commit attempts to provide the firsts steps to check this.