Add basic ClientVPN Service Checker
This commit adds a basic checker that consists in a set of REST calls
that makes the ClientVPN server retireve the config files and change
it's internal state to finally allow the docker calls.

Since the ClientVPN has "state" (Basically it does not allow any docker
related call to be run before the config files are there), this file
follow the order as expected.

Of course, the bad workflow was also tested and formal Test Suites
showing this will come eventually