Correct Service Maker Autogenerate GET Method with URL Params
When the service maker had to genereate a method from a client with URL
Params, the lambda function created was expecting params in *arg format.

Due to this, we can generate a generic function that successively
overrides the %s  previously detected and formatted. We can do this
by putting the args after the url string as list or tuple.

However there was a bug that was calling the str() methods to the value
of args. That conversion was converting a list of string into a string
that represents a list.