Add Blocking VPN Deployment
Before this commit, when the deploy method was called, if the status
was "ready", the VPN was deployed, if not, it created a thread that
was checking the status and was called when the status was ready,
without blocking the exectution.

Since we are deploying the system in SlipStream VMs, there are some
problems with that approach. First, our configuration and installation
scripts do not block nor wait for the VPN to be read. Second, doing
that may need some rework on the scripts to accomplish that goal.

For that reason, we implemented a deploy_blocking() call that blocks
the execution until the VPN is at "ready" status