Fix bug in service maker related to post with json.
The bug caused json attribute to be ignored in post requests.
The second 'if' clause was evaluated to false, so its 'else' clause
was executed, overriding func to one without json.

With this fix, only one clause is evaluated true, which results in
func being assigned only once. In case of post, assigned to a func
with json argument.